// src/app/interfaces/evento.ts
export interface Evento {
    documentName: string;
    documentDescription: string;
    eventType: string;
    eventStartDate: string;
    eventEndDate: string;
    eventCountryName: string;
    eventTerritoryName: string;
    eventTownName: string;
    eventLocation: string;
    eventLocationName: string;
    eventWhere: string;
    eventSponsor: string;
    eventRegistrationStartDate: string;
    eventRegistrationEndDate: string;
    eventOnLine: string;
    eventSearchDate1: string;
    eventSearchDate2: string;
    importance: string;
    profile: string;
    parentcorec: string;
    marks: string;
    templateType: string;
    productClub: string;
    visit: string;
    store: string;
    eventTimeTable: string;
    eventPrice: string;
    eventDuration: string;
    eventLanguages: string;
    eventImageUrl: string;
    eventSourceUrl: string;
    eventStatus: string;
    eventTheme: string;
    eventTargetGroup: string;
    latitudeLongitude: string;
    latwgs84: string;
    lonwgs84: string;
    placeName: string;
    address: string;
    municipality: string;
    municipalityCode: string;
    postalCode: string;
    territory: string;
    territoryCode: string;
    country: string;
    countryCode: string;
    phone: string;
    email: string;
    webpage: string;
    friendlyUrl: string;
    physicalUrl: string;
    dataXML: string;
    metadataXML: string;
    zipFile: string;
  } 